Redirect to default pages from home
